+++ R/NDFE.R
@@ -76,11 +76,14 @@ NDFE.fit <- function (t, C, A = 1, V, serie = "", k = log(0.01), verbose = TRUE,
   
   tryCatch({
     stopifnot(length(t) > 3)
-    fit <- nls(C ~ cbind(1, exp(k)*(A/V)*(2/sqrt(pi)*sqrt(t/exp(k))+exp(t/exp(k))*erfc(sqrt(t/exp(k)))-1)), 
+    fit <- withWarnings(nls(C ~ cbind(1, exp(k)*(A/V)*(2/sqrt(pi)*sqrt(t/exp(k))+exp(t/exp(k))*erfc(sqrt(t/exp(k)))-1)), 
                start= list(k=k), algorithm = "plinear",
-               control=nls.control(maxiter=maxiter, minFactor=1e-10))
-    
-    fitsum <- summary(fit)
+               control=nls.control(maxiter=maxiter, minFactor=1e-10)))
+    w <- if (is.null(fit$warnings)) "" else fit$warnings[[1]]$message
+    fit <- fit$value
+    fitsum <- withWarnings(summary(fit))
+    if (w == "") w <- if (is.null(fitsum$warnings)) "" else fitsum$warnings[[1]]$message
+    fitsum <- fitsum$value
     fitsumCoef <- fitsum$coef
     
     try({
@@ -98,12 +101,12 @@ NDFE.fit <- function (t, C, A = 1, V, serie = "", k = log(0.01), verbose = TRUE,
       AIC=AIC(fit),
       AICc=AICc(fit),
       RSE=fitsum$sigma,
-      diagnostics = "")
-    if (verbose) message(serie, ": NDFE fit successful")
+      diagnostics = w)
+    if (verbose) message(serie, if (w == "") ": NDFE fit successful" else ": NDFE fit warning")
     res
   },
   error = function(cond) {
-    if (verbose) message(serie, ": NDFE fit not successful")
+    if (verbose) message(serie, ": NDFE fit failed")
     list(
       f0 = NA_real_, 
       f0.se = NA_real_, 
@@ -114,27 +117,6 @@ NDFE.fit <- function (t, C, A = 1, V, serie = "", k = log(0.01), verbose = TRUE,
       AICc=NA_real_,
       RSE=NA_real_,
       diagnostics=cond$message)
-  },
-  warning = function(cond) {
-    fitsum <- summary(fit)
-    fitsumCoef <- fitsum$coef
-    try({
-      if (plot) {
-        curve(predict(fit, newdata = data.frame(t = x)), 
-              from = min(t), to = max(t), add = TRUE, col = "blue")
-      }}, silent = TRUE)
-    res <- list(
-      f0 = fitsumCoef[".lin2", "Estimate"], 
-      f0.se = fitsumCoef[".lin2", "Std. Error"], 
-      f0.p = fitsumCoef[".lin2", "Pr(>|t|)"], 
-      tau=exp(fitsumCoef["k", "Estimate"]),
-      C0=fitsumCoef[".lin1", "Estimate"],
-      AIC=AIC(fit),
-      AICc=AICc(fit),
-      RSE=fitsum$sigma,
-      diagnostics = cond$message)
-    if (verbose) message(serie, ": NDFE fit successful (warning)")
-    res
   }
   )  
 }